# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "BHUVISION \u2014 Agentic Earth Intelligence",
    "description": "An Interactive Vision-Language Assistant for Multimodal Remote Sensing Image Analysis through Text Queries (SIH26167, ISRO). Powered by 9 autonomous specialist agents.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/imagery/upload": {
      "post": {
        "tags": [
          "Imagery"
        ],
        "summary": "Upload Image",
        "description": "Upload a satellite image (GeoTIFF, JP2, PNG, JPEG) and parse geospatial metadata.",
        "operationId": "upload_image_api_imagery_upload_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_image_api_imagery_upload_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImageryUploadResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/imagery/{image_id}": {
      "get": {
        "tags": [
          "Imagery"
        ],
        "summary": "Get Imagery Metadata",
        "description": "Retrieve metadata for an uploaded or cached satellite image.",
        "operationId": "get_imagery_metadata_api_imagery__image_id__get",
        "parameters": [
          {
            "name": "image_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Image Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImageryMetadata"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BandInfo": {
        "properties": {
          "index": {
            "type": "integer",
            "title": "Index"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "dtype": {
            "type": "string",
            "title": "Dtype",
            "default": "float32"
          },
          "nodata": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nodata"
          },
          "min_value": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Min Value"
          },
          "max_value": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Value"
          }
        },
        "type": "object",
        "required": [
          "index"
        ],
        "title": "BandInfo",
        "description": "Information about a raster band."
      },
      "Body_upload_image_api_imagery_upload_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_image_api_imagery_upload_post"
      },
...
      "GeoMetadata": {
        "properties": {
          "crs": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Crs",
            "description": "Coordinate Reference System (e.g., EPSG:4326)"
          },
          "bounds": {
            "anyOf": [
              {
                "items": {
                  "type": "number"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bounds",
            "description": "[west, south, east, north] in CRS units"
          },
          "transform": {
            "anyOf": [
              {
                "items": {
                  "type": "number"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Transform",
            "description": "Affine transform coefficients"
          },
          "width": {
            "type": "integer",
            "title": "Width",
            "default": 0
          },
          "height": {
            "type": "integer",
            "title": "Height",
            "default": 0
          },
          "band_count": {
            "type": "integer",
            "title": "Band Count",
            "default": 0
          },
          "bands": {
            "items": {
              "$ref": "#/components/schemas/BandInfo"
            },
            "type": "array",
            "title": "Bands"
          },
          "resolution": {
            "anyOf": [
              {
                "prefixItems": [
                  {
                    "type": "number"
                  },
                  {
                    "type": "number"
                  }
                ],
                "type": "array",
                "maxItems": 2,
                "minItems": 2
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolution",
            "description": "Pixel size (x, y) in CRS units"
          }
        },
        "type": "object",
        "title": "GeoMetadata",
        "description": "Geospatial metadata extracted from imagery."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ImageFormat": {
        "type": "string",
        "enum": [
          "geotiff",
          "jp2",
          "png",
          "jpeg",
          "safe",
          "unknown"
        ],
        "title": "ImageFormat",
        "description": "Supported image formats."
      },
      "ImageryMetadata": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "description": "Unique imagery identifier"
          },
          "filename": {
            "type": "string",
            "title": "Filename"
          },
          "file_size_bytes": {
            "type": "integer",
            "title": "File Size Bytes",
            "default": 0
          },
          "format": {
            "$ref": "#/components/schemas/ImageFormat",
            "default": "unknown"
          },
          "sensor_type": {
            "$ref": "#/components/schemas/SensorType",
            "default": "unknown"
          },
          "acquisition_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Acquisition Date"
          },
          "satellite": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Satellite",
            "description": "e.g., Sentinel-2, Sentinel-1, Landsat-8"
          },
          "geo": {
            "$ref": "#/components/schemas/GeoMetadata"
          },
          "extra": {
            "additionalProperties": true,
            "type": "object",
            "title": "Extra"
          }
        },
        "type": "object",
        "required": [
          "id",
          "filename"
        ],
        "title": "ImageryMetadata",
        "description": "Complete metadata for a satellite image."
      },
      "ImageryUploadResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "filename": {
            "type": "string",
            "title": "Filename"
          },
          "metadata": {
            "$ref": "#/components/schemas/ImageryMetadata"
          },
          "preview_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preview Url"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          }
        },
        "type": "object",
        "required": [
          "id",
          "filename",
          "metadata"
        ],
        "title": "ImageryUploadResponse",
        "description": "Response after uploading imagery."
      },
...
      "SensorType": {
        "type": "string",
        "enum": [
          "optical",
          "sar",
          "unknown"
        ],
        "title": "SensorType",
        "description": "Type of satellite sensor."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}